Quick and dirty MSF RPC daemon status notification
This may not be the ultimate way to do this, but seems to work.  Basically, I've been getting
tripped up on what's the status of the connection to the MSF daemon if I missed the toast.
This lets me see at a glance what the current status is.

It's on by default but you can turn it off in preferences.  Eventually we might want to replace
the toast with a material-esque Snackbar.