Speed up port protocol resolution.
Workaround for #332. Instead of loading services when we find a new open port, do it when we enter in the prrt scanner, and in a thread. So when a new open port is discovered, it has already loaded ~2000 ports.

Also avoid to create new objects on every iteration of the loops.