Quick (not by any means final) materialization of buttons w/fabs/colors.
Not everything here is done as well as it could be, but it's meant only as
a step in the right direction.  It's possible some stuff was inadvertantly
broken.  Could use animation and stuff, and probably shouldn't hard-code in
so much color and the like.  A few buttons and such may have been missed,
but I think it's by-and-large better than the plain grey buttons that were
there before.