manage the navigation drawer from AbstractDrawerActivity.
created a progress dialog that will show up only if
a certain amount of time has been elapsed and disappear if
has been shown for at least another certain time.

show such progress dialog when changing interface
( stopping and starting network radar may require some second ).

another way is to asynchronous stopping the network-radar but
we should deregister receivers and we lost the STARTED/STOPPED
intents.

in a far future we will support more running instances of
network-radar, so you can be notified ( say, with a small icon over the interface name )
that new hosts appearead on interfaces that you are not watching.